Enhance file operations and debugging for 64-bit support
- Updated function signatures in file_operations.h, file_list.h, and debug.h to use 'unsigned long long' for size parameters.
- Refactored file size handling in file_operations.c to utilize 64-bit operations for larger files.
- Modified debug messages to reflect changes in data types for consistency and clarity.
- Adjusted file_list.c to accommodate new data types in file entry structures and related functions.
- Ensured all relevant functions in main.c and batch.c are updated to handle 64-bit sizes correctly.

+++ include/utils/file_list.h
@@ -12,9 +12,9 @@ typedef struct file_entry
 {
     char input_filename[FILE_LIST_MAX_FILENAME];
     char output_filename[FILE_LIST_MAX_FILENAME];
-    unsigned long sequence_num;
-    unsigned long original_size;
-    unsigned long processed_size;
+    unsigned long long sequence_num;
+    unsigned long long original_size;
+    unsigned long long processed_size;
     struct file_entry *next;
 } file_entry_t;
 
@@ -23,8 +23,8 @@ typedef struct
 {
     file_entry_t *head;
     file_entry_t *tail;
-    unsigned long count;
-    unsigned long next_sequence_num;
+    unsigned long long count;
+    unsigned long long next_sequence_num;
 } file_list_t;
 
 /* Initialise file list */
@@ -32,15 +32,15 @@ int file_list_init(file_list_t *list);
 
 /* Add file to list */
 int file_list_add(file_list_t *list, const char *input_filename,
-                  const char *output_filename, unsigned long original_size,
-                  unsigned long processed_size);
+                  const char *output_filename, unsigned long long original_size,
+                  unsigned long long processed_size);
 
 /* Find file in list by name */
 file_entry_t *file_list_find(file_list_t *list, const char *filename);
 
 /* Get most recent files */
-unsigned long file_list_get_recent(file_list_t *list, unsigned long count,
-                                   file_entry_t **result);
+unsigned long long file_list_get_recent(file_list_t *list, unsigned long long count,
+                                        file_entry_t **result);
 
 /* Save list to file */
 int file_list_save(file_list_t *list, const char *filename);
